Move `modular::bingcd` content to `uint::bingcd`
