+++ src/uint.rs
@@ -26,6 +26,7 @@ mod macros;
 
 mod add;
 mod add_mod;
+mod bingcd;
 mod bit_and;
 mod bit_not;
 mod bit_or;
